Initialized bodyparser. User can now create widget items - AddWidget form component sends data to the Database component, from here, data gets sent to the backend using axios, backend takes the data inside the route and creates a widget row in the database, successful data gets sent back to frontend, that data gets dispatched to rootReducer, where the Redux state is updated. Used Sequelize.NUMERIC so that decimals can be used

+++ server/models/widget.js
@@ -18,7 +18,7 @@ module.exports = function (sequelize, Sequelize) {
     },
 
     aluminum: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.NUMERIC,
       allowNull: false,
       validate: {
         notEmpty: true,
@@ -26,7 +26,7 @@ module.exports = function (sequelize, Sequelize) {
     },
 
     cold_rolled_steel: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.NUMERIC,
       allowNull: false,
       validate: {
         notEmpty: true,
@@ -34,7 +34,7 @@ module.exports = function (sequelize, Sequelize) {
     },
 
     galvanneal: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.NUMERIC,
       allowNull: false,
       validate: {
         notEmpty: true,
@@ -42,7 +42,7 @@ module.exports = function (sequelize, Sequelize) {
     },
 
     glass: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.NUMERIC,
       allowNull: false,
       validate: {
         notEmpty: true,
@@ -50,7 +50,7 @@ module.exports = function (sequelize, Sequelize) {
     },
 
     stainless_steel: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.NUMERIC,
       allowNull: false,
       validate: {
         notEmpty: true,
